requirements.txt
# Requires Python >= 3.13
# Install: uv pip install -r requirements.txt
#          pip install -r requirements.txt

# LLM core
openai>=2.0
pydantic>=2.7
pyyaml>=6.0
httpx>=0.27
python-dotenv>=1.0

# Math runtime for run_python tool
sympy>=1.13
numpy>=1.26
scipy>=1.13